Introduce new formula auto complete feature
When we start to type a formula in a cell eg: =S
We will see the possible function names in a popup.
If we move with "ArrowDown and ArrowUp" and enter, or direcly select a
function with mouse, function name will be automatically inserted into
cell. Then we will see the usage details of the selected function in
another dialog.

Signed-off-by: Gülşah Köse <gulsah.kose@collabora.com>
Change-Id: Ic06ca0d904770e95bb85ccdf9a73f60b7d5c5eb0