Toolbar-down (status bar): Fix hover status
For all status bar buttons
- Fix missing pointer
- Also, at leat for now, don't add border and don't add different
background. This is a status bar and adding those contributes to the
overall bulkiness. In the future we could perhaps remove the vertical
padding of the status bar and introduce back the background on
over (and remove the border radius). This would make the status bar
slimmer and slicker (so maybe we can turn its visibility on by default)

Signed-off-by: Pedro Pinto Silva <pedro.silva@collabora.com>
Change-Id: I2b01d1d6c35ff2b3f487de0417170398d587a565