fix: fix UI preference inconsistencies
Add helper functions for fetching preferences from localstorage and
ui_defaults. By having these in unified helpers, we can be certain that
we're not making differences in the priority or use of different stores
which can cause UI inconsistency.

BREAKING-CHANGE: This changes the naming of many of the preferences in localStorage, effectively resetting lots of UI preferences

Adapt patch to LibreOffice Online.

Signed-off-by: Skyler Grey <skyler.grey@collabora.com>
Change-Id: I4cc9cc885aa8cd0b40646629cd73bef236963a94
Fixes: CollaboraOnline/online#9166