fix: allow changing compact to tabbed mode
This is a fix for a regression from
I4cc9cc885aa8cd0b40646629cd73bef236963a94

Compact/tabbed mode was saved on a per-app basis, but read on a global
basis. This meant that the read preference was never changed, causing it
to be impossible to switch out of compact mode once you had entered it.

The fix is to change back to the previous behavior of saving and reading
on a global basis

Refs: CollaboraOnline/Online#9181
Signed-off-by: Skyler Grey <skyler.grey@collabora.com>
Change-Id: Ie9d72d6b7e10e209f10acf662535c866e06a98de