feat: record the peformance of CPU,Network,Latency
The peformance of the CPU is measured by the run time in ms, Network by
the incoming and outgoing bandwith in kB/s and latency is split into
buckets of ping and tiles in ms. The data is stored in seperate CSV
files which will be updated each unit test allowing the peformance data
to be in a form which can be re-used, graphed or at-least monitored in
the future.

Signed-off-by: elliotfreebairn <elliot.freebairn@collabora.com>
Change-Id: I415d414c832a00d2b55815df89f28a2d836a1290