All comment and white space changes.
It's not possible to compile with "--enable-bundle" for debug.

If one needs to test the bundle, uglify.js complains about the line endings of "sourcemap" indicators.
Sourcemap comments are automatically added to the files and they don't use a new line after the filepath.
This causes comments starting with /* chars to be on the same line of source map paths.
So the rest of the actual comments are seen like code lines.
These changes ensure that files compile with the "--enable-bundle" flag.

Adapt patch to LibreOffice Online.

Signed-off-by: Gökay Şatır <gokaysatir@collabora.com>
Change-Id: Ia5ba5b3ea503efd7005a5821f6e8028867025e42