feat: New peak memory, cpu and network measuring
Instead of measuring the full pss value, only recording dirty pages to get a more solid numbers. Removed measuring CPU Usage from /proc and using SysStopWatch to measure the elapsed time of the CPU for the test. Also switched out recording the rate at which bytes have been recieved and sent to the total of bytes.

Signed-off-by: elliotfreebairn <elliot.freebairn@collabora.com>
Change-Id: If74d883ba3f75a42bf0b78d0757d8f7bfe747237